:truck: api: move ComplementaryCertificationCourseResult model to src
