+++ api/db/database-builder/factory/build-complementary-certification-course-result.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import { ComplementaryCertificationCourseResult } from '../../../lib/domain/models/ComplementaryCertificationCourseResult.js';
+import { ComplementaryCertificationCourseResult } from '../../../src/certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
 import { databaseBuffer } from '../database-buffer.js';
 import { buildBadge } from './build-badge.js';
 import { buildCertificationCourse } from './build-certification-course.js';
