+++ api/lib/domain/models/CertificationResult.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
 import { CompetenceMark } from '../../../src/certification/results/domain/models/CompetenceMark.js';
+import { ComplementaryCertificationCourseResult } from '../../../src/certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
 import { JuryComment, JuryCommentContexts } from '../../../src/certification/shared/domain/models/JuryComment.js';
-import { ComplementaryCertificationCourseResult } from './ComplementaryCertificationCourseResult.js';
 
 /**
  * @readonly
