+++ api/tests/unit/domain/events/handle-complementary-certifications-scoring_test.js
@@ -1,7 +1,7 @@
 import range from 'lodash/range.js';
 
 import { _forTestOnly } from '../../../../lib/domain/events/index.js';
-import { ComplementaryCertificationCourseResult } from '../../../../lib/domain/models/ComplementaryCertificationCourseResult.js';
+import { ComplementaryCertificationCourseResult } from '../../../../src/certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
 import { config as settings } from '../../../../src/shared/config.js';
 import { status as assessmentResultStatuses } from '../../../../src/shared/domain/models/AssessmentResult.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
