+++ api/tests/unit/domain/models/ComplementaryCertificationCourseWithResults_test.js
@@ -1,5 +1,5 @@
-import { sources } from '../../../../lib/domain/models/ComplementaryCertificationCourseResult.js';
 import { ComplementaryCertificationCourseWithResults } from '../../../../lib/domain/models/ComplementaryCertificationCourseWithResults.js';
+import { sources } from '../../../../src/certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
 import { expect } from '../../../test-helper.js';
 
 describe('Unit | Domain | Models | ComplementaryCertificationCourseWithResults', function () {
